Extract OnCallShift to models.py and dedupe date parsing
Two more follow-ups from the post-implementation retrospective:

3. The JSM command was using a deferred import to dodge a circular
   dependency: jsm.py imported OnCallShift from main, while main needed
   to import from jsm. Move OnCallShift to a new src/minuto/models.py
   so both clients import it from there. Lifts the deferred import in
   calculate_from_jsm to module level. `from minuto.main import
   OnCallShift` keeps working through the new top-level re-export, so
   no test or external consumer changes.

4. The ~10-line block that parses --start-date / --end-date and
   defaults a date-only end to 23:59:59 was duplicated between
   calculate_from_opsgenie and calculate_from_jsm. Extract it to a
   `_parse_window` helper. Net -18 lines in main.py.

Verified: 26/26 tests pass; live opsgenie + jsm commands still produce
bit-identical CSVs for the back-window case; bad-date error path still
exits 1 with the readable message.
